Moved the commons from the angular app only into the electron app. This way we can have them in one place only and use them accordingly
